requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
django-environ==0.11.2
Faker==20.1.0
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
django-cors-headers==4.3.1

# Development dependencies
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0
factory-boy==3.3.0 